moved table component from InstancesListTable to own component w/ burnettk
